+++ CHANGELOG.md
@@ -40,6 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactor ItemCard and InventorySection to render forms and additional components.
 
 ### Fixed
+
+- Update AddShoppingListItemForm to use dynamic ID for quantity input
 - Correctly maps ExpiresOn from destructured item within ItemCard and removes title, quantity, expirationDate in favor of their destructured values
 - Fixes mock item in test to pass in ExpiresOn instead of ExpirationDate